The two lines of work are orthogonal and every conflict was additive: the
enzyme presets say where cleavage may occur, digestion specificity says
how many termini must respect it. BuildParams carries both cleavageRegex
and digestSpecificity, and a config records cleavage_regex alongside
specificity.

Three resolutions worth naming. types.ts and config.ts keep both fields
rather than either. App.tsx keeps his defaults-merge when recalling a
stored run -- older snapshots carry neither field -- with the
downloadspeclib branch restored beside it.

digest_sequence gained a required specificity argument and a third return
value, so test_cleavage_presets.jl passes "full" explicitly. Those tests
assert what a cleavage rule identifies as a site, which is the same
question whichever specificity is layered on top.

+++ src/structs/protein_inference_types.jl
@@ -72,18 +72,18 @@ Base.hash(k::PeptideKey, h::UInt) = hash((k.sequence, k.is_target, k.entrap_id),
 Statistical features of a protein group used for scoring and FDR control.
 
 # Fields
-- `n_peptides::UInt16`: Number of unique peptides observed
-- `n_possible_peptides::UInt16`: Total possible peptides from in silico digestion
+- `n_peptides::UInt32`: Number of unique peptides observed
+- `n_possible_peptides::UInt32`: Total possible peptides from in silico digestion
 - `peptide_coverage::Float32`: Fraction of possible peptides observed
-- `total_peptide_length::UInt16`: Sum of lengths of all observed peptides
+- `total_peptide_length::UInt32`: Sum of lengths of all observed peptides
 - `log_n_possible_peptides::Float32`: Log of n_possible_peptides for regression
 - `log_binom_coeff::Float32`: Log binomial coefficient for peptide sampling
 """
 struct ProteinFeatures
-    n_peptides::UInt16
-    n_possible_peptides::UInt16
+    n_peptides::UInt32
+    n_possible_peptides::UInt32
     peptide_coverage::Float32
-    total_peptide_length::UInt16
+    total_peptide_length::UInt32
     log_n_possible_peptides::Float32
     log_binom_coeff::Float32
 end
